made Models folder and moved models into added Weather.swift
