+++ iOS-training/LayoutPrototype.swift
@@ -100,31 +100,3 @@ struct LayoutPrototype: View {
 #Preview {
     LayoutPrototype()
 }
-
-enum Weather: String, Decodable {
-    case sunny = "sunny"
-    case cloudy = "cloudy"
-    case rainy = "rainy"
-
-    var icon: Image {
-        switch self {
-        case .sunny:
-            return Image(.iconsun)
-        case .cloudy:
-            return Image(.iconclouds)
-        case .rainy:
-            return Image(.iconumbrella)
-        }
-    }
-
-    var color: Color {
-        switch self {
-        case .sunny:
-            return .red
-        case .cloudy:
-            return .gray
-        case .rainy:
-            return .blue
-        }
-    }
-}
